backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-lru==2.0.4
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deep-translator==1.11.4
edge-tts==7.2.7
fastapi==0.129.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
MarkupSafe==3.0.3
multidict==6.7.1
nest-asyncio==1.6.0
passlib==1.7.4
pg8000==1.31.5
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
scramp==1.4.8
six==1.17.0
soupsieve==2.8.3
SpeechRecognition==3.14.5
SQLAlchemy==2.0.46
starlette==0.52.1
tabulate==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.5
yarl==1.22.0
APScheduler==3.10.4
scikit-learn
pandas==2.2.3
prophet==1.1.6
psycopg2-binary
gunicorn==23.0.0
groq==0.15.0
openai-whisper==20231117
